fix: decouple VS Code imports from agent, tools, and event bus modules
Remove all `vscode` and `@frontend/` imports from VS Code-free zones
(`src/agent/`, `src/tools/`, `src/eventBus/`) as specified in CLAUDE.md.

- **executeAgent.ts**: Replace 5 vscode API calls with event bus emissions
  (`requestShowError`, `requestShowAgentNotification`) and an injectable
  `setShowProgressView()` callback for the awaited progress view command.

- **toolEditApproval.ts**: Extract all VS Code UI code (diff editor, temp
  files, tab management) into `src/frontend/approval/nativeToolEditApproval.ts`.
  The pure module keeps business logic, queue management, bypass state, and
  diff computation. The native handler registers via `setToolEditApprovalHandler`.

- **bashApproval.ts**: Replace `@frontend/system/commandUtils` import with
  `bus.emit('requestExecuteCommand', ...)` event.

- **VscodeIntegration.ts**: Move from `src/tools/lean/` to `src/frontend/lean/`.
  Create injectable `LeanVscodeServices` interface in `src/tools/lean/` so
  `LspTools.ts` accesses VS Code integration without direct imports.

- **LspTools.ts**: Replace direct VscodeIntegration import and
  `@frontend/vscode/vscodeDiagnostics` import with injectable services and
  shared types from `leanTypes.ts`.

New event bus events: `requestShowError`, `requestShowAgentNotification`,
`requestExecuteCommand` — all handled in `agentEventListeners.ts`.
