+++ src/frontend/events/agentEventListeners.ts
@@ -93,6 +93,16 @@ async function handleRequestEnsureProgressView(
   }
 }
 
+function handleRequestExecuteCommand(
+  payload: ProgressEventPayloads['requestExecuteCommand'],
+): void {
+  void vscode.commands.executeCommand(
+    payload.command,
+    ...(payload.args ?? []),
+  );
+}
+
+
 /**
  * Register all agent→frontend event listeners.
  * Returns a Disposable that cleans up all subscriptions.
@@ -114,6 +124,7 @@ export function registerAgentEventListeners(): vscode.Disposable {
     (payload) => void handleRequestEnsureProgressView(payload),
     { signal },
   );
+  bus.on('requestExecuteCommand', handleRequestExecuteCommand, { signal });
 
   return new vscode.Disposable(() => controller.abort());
 }
